Moved ...::detail::zlib_compressor to ...::zlib::compressor
* compressor is now a public component, which is expected
  (eventually) to be plugged into the codec.
