+++ include/telnetpp/options/mccp/zlib/compressor.hpp
@@ -1,32 +1,38 @@
 #pragma once
 
-#include "telnetpp/core.hpp"
+#include "telnetpp/options/mccp/compressor.hpp"
+#include <memory>
 
-namespace telnetpp { namespace options { namespace mccp { namespace detail {
+namespace telnetpp { namespace options { namespace mccp { namespace zlib {
 
-//* =========================================================================
-/// \brief Represents an object that can compress arbitrary byte sequences.
-//* =========================================================================
-class compressor
+class TELNETPP_EXPORT compressor : public telnetpp::options::mccp::compressor
 {
 public :
     //* =====================================================================
+    /// \brief Constructor
+    //* =====================================================================
+    compressor();
+    
+    //* =====================================================================
     /// \brief Destructor
     //* =====================================================================
-    virtual ~compressor() {}
+    ~compressor() override;
     
     //* =====================================================================
     /// \brief Compress the given byte sequence and return the compressed
     /// sequence.
     //* =====================================================================
-    virtual telnetpp::u8stream compress(
-        telnetpp::u8stream const &sequence) = 0;
-
+    telnetpp::u8stream compress(telnetpp::u8stream const &sequence) override;
+    
     //* =====================================================================
     /// \brief Ends the compression stream.  Compressing further will cause
     /// the stream to restart.
     //* =====================================================================
-    virtual telnetpp::u8stream end_compression() = 0;
+    telnetpp::u8stream end_compression() override;
+
+private :
+    struct impl;
+    std::unique_ptr<impl> pimpl_;
 };
 
 }}}}
